refatorar tipo Sprite e função DesenharSprite para outro arquivo
