+++ sprite.c
@@ -1,0 +1,5 @@
+#include "sprite.h"
+
+void DesenharSprite(Sprite s) {
+  DrawTexturePro(*s.textura, s.src, s.dst, (Vector2){ 0.0, 0.0 }, 0.0, WHITE);
+}
